Fix error #rgb on wp piker when save a example
Wp always inserts # in front of the text that does not present it. This
conflicts with alpha color that uses rgb and rgba.

The function that removes # in front of rgb has been implemented, in
addition to rgba, also for rgbs that were loaded by the example.

Additionally, the function was moved from the original position of the
"Grad_Ncolor_change" function to a new function
"Grad_wpcolorpiker_error_autoHEX".

The new function is loaded as before as soon as you load the js of the
GradGeneratorSimply.js file